+++ symbio/src/plugins/model/protocols/openai_chat.rs
@@ -10,6 +10,7 @@ use super::super::types::CapabilityMeta;
 use super::partial_json::{FieldPath, JsonLineExtractor, PartialJsonSink, StrAction};
 use super::{sse_data, ModelProtocol, MODEL_PROTOCOL_OPENAI_CHAT};
 use crate::symbio_core::sse::PartialLineExtractor;
+use crate::symbio_core::tool_name::to_wire;
 use crate::symbio_core::{
     get_http_client, FinishReason, InvokeRequest, PluginError, ProtocolEvent, SseLineParser, Usage,
 };
@@ -71,7 +72,7 @@ impl ModelProtocol for OpenaiChatProtocol {
                     json!({
                         "type": "function",
                         "function": {
-                            "name": t.name.replace("/", "__"),
+                            "name": to_wire(&t.name),
                             "description": t.description_for_llm(),
                             "parameters": t.input_schema
                         }