+++ symbio/src/plugins/model/protocols/openai_responses.rs
@@ -10,6 +10,7 @@ use super::super::types::{CapabilityMeta, ContentPart, MessageContent, MessageRo
 use super::partial_json::{FieldPath, JsonLineExtractor, PartialJsonSink, StrAction};
 use super::{sse_data, ModelProtocol, MODEL_PROTOCOL_OPENAI_RESPONSES};
 use crate::symbio_core::sse::PartialLineExtractor;
+use crate::symbio_core::tool_name::to_wire;
 use crate::symbio_core::{
     get_http_client, FinishReason, InvokeRequest, PluginError, ProtocolEvent, SseLineParser, Usage,
 };
@@ -187,7 +188,7 @@ impl ModelProtocol for OpenaiResponsesProtocol {
                 .iter()
                 .map(|t| json!({
                     "type": "function",
-                    "name": t.name.replace("/", "__"),
+                    "name": to_wire(&t.name),
                     "description": t.description_for_llm(),
                     "parameters": t.input_schema
                 }))